v1.18.5 (Chrome only)
Just released v1.18.5 for Chrome. No time for a Firefox update yet, but
I'll get  to it soon(ish). There are still some issues. See
http://goo.gl/6pASR for details or to report any problems. Sorry it's
not perfect yet, but I figure it's better to get something out sooner
rather than later. More to come ....

As many of you have noticed, my old site is gone due to the hassle of
trying to maintain it. Everything for this project is now hosted on
GitHub.

As always, feel free to buy me a cup of coffee via PayPal
(http://goo.gl/8p5ru), and special thanks to all of you who have kept me
caffeinated over the years and supported this project. ♥